refactor: change value type in TransferRequest from Double to BigDecimal for improved precision

+++ src/main/java/com/lucascosta/desafiopagamento/adapters/inbound/dto/TransferRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.Digits;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public record TransferRequest(
 
         @NotNull(message = "O ID do pagador é obrigatório")
@@ -15,6 +17,6 @@ public record TransferRequest(
         @Positive(message = "O valor da transferência deve ser positivo")
         @Digits(integer = 10, fraction = 2, message = "O valor da transferência deve ter no máximo 10 dígitos inteiros e 2 decimais")
         @NotNull(message = "O valor da transferência é obrigatório")
-        Double value
+        BigDecimal value
 ) {
 }
